+++ django_amazon_ses.py
@@ -39,6 +39,8 @@ class EmailBackend(BaseEmailBackend):
         secret_access_key = getattr(settings, 'AWS_SES_SECRET_ACCESS_KEY',
                                     secret_access_key)
         region_name = getattr(settings, 'AWS_SES_REGION', region_name)
+        self.configuration_set_name = getattr(
+            settings, 'AWS_SES_CONFIGURATION_SET_NAME', None)
 
         self.conn = boto3.client(
             'ses',
@@ -92,13 +94,18 @@ class EmailBackend(BaseEmailBackend):
         message = email_message.message().as_bytes(linesep='\r\n')
 
         try:
-            result = self.conn.send_raw_email(
-                Source=from_email,
-                Destinations=recipients,
-                RawMessage={
+            kwargs = {
+                'Source': from_email,
+                'Destinations': recipients,
+                'RawMessage': {
                     'Data': message
-                }
-            )
+                },
+            }
+
+            if self.configuration_set_name is not None:
+                kwargs['ConfigurationSetName'] = self.configuration_set_name
+
+            result = self.conn.send_raw_email(**kwargs)
             message_id = result['MessageId']
             post_send.send(
                 self.__class__,